Catch up to build net on linux.
Added a linux net_resources target for the resource itself, and an
action for using tld_cleanup on the generated .dat file linux uses.

Adds generation of the current SCons pattern for building resources.

Generalized how we build different types of things into a table to
make things a little more modular.

Update for configurations now being a dict indexed by name.

Required a Linux version of icudt38l_dat.s without the .const pseudo-op
that the stock GNU assembler doesn't understand, and without the leading
underscore on the symbols.