Regression test and fix for http://crbug.com/122588.
The xcode generator can't handle if 'dependencies' contains the
same target twice, as it happens if the dependencies list contains
both 'a.gyp:a' and 'a.gyp:*'.

Instead of changing the xcode generator, let input.py unify dependency
lists before passing them on to the generator.

I ran build/gyp_chromium a few times with and without this change and
couldn't see a difference in running time.

BUG=chromium:122588
Review URL: https://chromiumcodereview.appspot.com/10027005