Preserve the mtime of solibs whose API hasn't changed, and enable ninja restat.
This makes it so that changing component-internal implementation details only
causes the component lib to be rebuilt, not downstream dependencies.

For example touching a content/ .cc file and rebuilding "chrome" causes the .cc
file to be recompiled and the new .o to be relinked into lib/libcontent.so, but
the rebuild ends there, since the chrome binary itself doesn't need to be
relinked.

TEST=The second ninja command in this pipeline doesn't relink chrome: ../ninja chrome && touch content/common/gpu/client/webgraphicscontext3d_command_buffer_impl.cc && ../ninja chrome

Patch from Ami Fischman <fischman@chromium.org>!
Review URL: http://codereview.chromium.org/10578031/