Parallelize ninja generation when GYP_PARALLEL=1.
This is a really easy approach, parallelizing by configs (Debug/Release). 
It's probably not worth parallelizing this step much more because there's 
some overhead and the total time of this step isn't that large.

This gives a huge speedup on Windows, and it's about even on a fast
Mac/Linux box.

Review URL: http://codereview.chromium.org/11026061/