win msvs: make ordering match .gyp order
The order of .obj on the link command line determines the order in the
final binary in the absence of other need to re-order. Make them
consistent across the windows generators by making them match the order
in the .gyp file, and add a test for this behaviour. This was already
ninja's behaviour.

R=thakis@chromium.org
BUG=chromium:326030

Review URL: https://codereview.chromium.org/107293004