Reland "msvs/ninja win: Fix support for ImageHasSafeExceptionHandlers"
This commit fixes the x64 issue and adds a test.  The underlying issue was that
the msvs default value for ImageHasSafeExceptionHandlers is None for x64
targets.  In order to add the x64 test, we need to update the ninja generator
to use ml64.exe instead of ml.exe when building x64 targets.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/861273003